Adding sv namespace to default namespace list
This change requires https://github.com/phpcr/phpcr/pull/25 to be
merged and the submodule updated to include the change.

This fixes part of the 07_Export tests (for method Session::exportSystemView).
The Session::exportDocumentView test is still failing because the
'test' namespace is not defined in the exported XML data.